requirements.txt
langchain==1.2.3
langchain-openai==1.1.7
openai==2.15.0
langchain-community==0.4.1
langchain-core==1.2.6
langchain-classic==1.0.1
python-dotenv==1.2.1
faiss-cpu==1.13.2
pypdf==6.5.0
beautifulsoup4==4.14.3
pydantic-settings==2.12.0
langchain-chroma==1.1.0
python-docx==1.2.0
pypdf==6.7.0
boto3==1.40.38